+++ src/App.js
@@ -9,7 +9,7 @@ import ManagerProjectPage from './Pages/ProjectPages/ManagerProjectPage';
 import { Route, Routes, useNavigate } from "react-router-dom";
 import { EmployeeProvider } from './Components/EmployeeContext/EmployeeContext';
 import StaffFeedbackPage from "./Pages/FeedbackPages/StaffFeedbackPage";
-import HRBookMeals from './Pages/MealPages/HRBookMeals';
+import HRPlanMeals from './Pages/MealPages/HRPlanMeals';
 import ChooseTime from './Pages/TimePages/ChooseTime';
 import Timer from './Pages/TimePages/Timer';
 
@@ -28,7 +28,7 @@ function App() {
           <Route path="/StaffProjectPage" element={<StaffProjectPage navigate={navigate}/>} />
           <Route path="/ManagerProjectPage" element={<ManagerProjectPage/>} />
           <Route path="/StaffFeedbackPage" element={<StaffFeedbackPage/>} />
-          <Route path="/HRBookMeals" element={<HRBookMeals/>} />
+          <Route path="/HRBookMeals" element={<HRPlanMeals/>} />
           <Route path="/ChooseTime" element={<ChooseTime/>} />
           <Route path="/Timer" element={<Timer/>} />
         </Routes>
