--- src/Pages/MealPages/HRBookMeals.css
@@ -1,71 +1,0 @@
-
-.book-meals-wrapper{
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    margin-top: 10%;
-    margin-left: 10%;
-    background-color: #FF6D00;
-    height: 60%;
-    width: 80%;
-    padding: 30px;
-    border-radius: 5px;
-    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.4);
-
-
-}
-.book-meals-wrapper h2{
-    justify-self: center;
-    align-self: center;
-}
-.book-entry{
-    display: flex;
-    flex-direction: column;
-}
-.book-entry input{
-    font-family:  Roboto, sans-serif;
-    height: 30px;
-    width: 400px;
-    border-radius: 5px;
-    border: 2px solid lightgray;
-}
-.book-entry input:focus{
-    background-color: whitesmoke;
-    outline: none;
-}
-.book-entry textarea{
-    font-family:  Roboto, sans-serif;
-    resize: none;
-    height: 65px;
-    border-radius: 5px;
-    border: 2px solid lightgray;
-}
-.book-entry textarea:focus{
-    outline: none;
-}
-.book-entry p{
-    margin-bottom: 5px;
-}
-.book-meals-button{
-        padding: 10px 20px;
-        border-radius: 5px;
-        border: none;
-        background-color: black;
-        color: #fff;
-        font-size: 16px;
-        cursor: pointer;
-        transition: background-color 0.3s;
-        margin-top: 10px;
-        justify-self: center;
-        align-self: center;
-}
-.book-meals-button:hover{
-    background-color: #383838;
-}
-
-.book-meals-wrapper .errorLabel {
-    color: rgb(255, 255, 255);
-    font-size: 13px;
-    margin-top: 10px;
-}
