+++ src/Pages/MealPages/HRPlanMeals.css
@@ -1,0 +1,86 @@
+.display-meals-wrapper{
+    font-family:  Roboto, sans-serif;
+    border: none;
+    background-color: #FF6D00;
+    border-radius: 5px;
+    padding-bottom: 30px ;
+    height: auto;
+    width: 80%;
+    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.4);
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    flex-direction: column;
+    margin: 10% auto;
+}
+
+.book-meals-wrapper{
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    margin-top: 10%;
+    margin-left: 8%;
+    background-color: #FF6D00;
+    height: 60%;
+    width: 70%;
+    padding: 30px;
+    border-radius: 5px;
+    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.4);
+
+
+}
+.book-meals-wrapper h2{
+    justify-self: center;
+    align-self: center;
+}
+.book-entry{
+    display: flex;
+    flex-direction: column;
+}
+.book-entry input{
+    font-family:  Roboto, sans-serif;
+    height: 30px;
+    width: 400px;
+    border-radius: 5px;
+    border: 2px solid lightgray;
+}
+.book-entry input:focus{
+    background-color: whitesmoke;
+    outline: none;
+}
+.book-entry textarea{
+    font-family:  Roboto, sans-serif;
+    resize: none;
+    height: 65px;
+    border-radius: 5px;
+    border: 2px solid lightgray;
+}
+.book-entry textarea:focus{
+    outline: none;
+}
+.book-entry p{
+    margin-bottom: 5px;
+}
+.book-meals-button{
+    padding: 10px 20px;
+    border-radius: 5px;
+    border: none;
+    background-color: black;
+    color: #fff;
+    font-size: 16px;
+    cursor: pointer;
+    transition: background-color 0.3s;
+    margin-top: 10px;
+    justify-self: center;
+    align-self: center;
+}
+.book-meals-button:hover{
+    background-color: #383838;
+}
+
+.book-meals-wrapper .errorLabel {
+    color: rgb(255, 255, 255);
+    font-size: 13px;
+    margin-top: 10px;
+}
